ui: rename SimulatedNodeStatus to NodeHistory
Release note: None
