+++ pkg/ui/ccl/src/views/clusterviz/containers/map/nodeSimulator.tsx
@@ -10,7 +10,6 @@ import React from "react";
 import { connect } from "react-redux";
 import * as d3 from "d3";
 
-import { NanoToMilli } from "src/util/convert";
 import { refreshNodes, refreshLiveness, refreshLocations } from "src/redux/apiReducers";
 import { selectLocalityTree, LocalityTier, LocalityTree } from "src/redux/localities";
 import { selectLocationsRequestStatus, selectLocationTree, LocationTree } from "src/redux/locations";
@@ -22,77 +21,14 @@ import {
   livenessStatusByNodeIDSelector,
   LivenessStatus,
 } from "src/redux/nodes";
-import { NodeStatus$Properties } from "src/util/proto";
 import { AdminUIState } from "src/redux/state";
 import Loading from "src/views/shared/components/loading";
 
 import { ZoomTransformer } from "./zoom";
 import { ModalLocalitiesView } from "./modalLocalities";
+import { NodeHistory } from "./nodeHistory";
 
 import spinner from "assets/spinner.gif";
-
-// SimulatedNodeStatus augments a node status from the redux state with
-// additional simulated information in order to facilitate testing of the
-// cluster visualization.
-//
-// TODO(mrtracy): This layer is a temporary measure needed for testing during
-// initial development. The simulator should be removed before official release,
-// with the simulated data coming from real sources.
-//
-// + HACK: Maintains old versions of each node status history, allowing for the
-//   computation of instantaneous rates without querying time series. This
-//   functionality should be moved to the reducer.
-// + Simulates long/lat coordinates for each node. This will eventually be
-//   provided by the backend via a system table.
-// + Simulates locality tiers for each node. These are already available from
-//   the nodes, but the simulated localities allow testing more complicated
-//   layouts.
-export class SimulatedNodeStatus {
-  // "Client Activity" is a generic measurement present on each locality in
-  // the current prototype design.
-  // Currently, it is the to number SQL operations executed per second,
-  // computed from the previous two node statuses.
-  clientActivityRate: number;
-  private statusHistory: NodeStatus$Properties[];
-  private maxHistory = 2;
-
-  constructor(initialStatus: NodeStatus$Properties) {
-    this.statusHistory = [initialStatus];
-    this.computeClientActivityRate();
-  }
-
-  update(nextStatus: NodeStatus$Properties) {
-    if (this.statusHistory[0].updated_at.lessThan(nextStatus.updated_at)) {
-      this.statusHistory.unshift(nextStatus);
-      if (this.statusHistory.length > this.maxHistory) {
-        this.statusHistory.pop();
-      }
-
-      this.computeClientActivityRate();
-    }
-  }
-
-  id() {
-    return this.statusHistory[0].desc.node_id;
-  }
-
-  latest() {
-    return this.statusHistory[0];
-  }
-
-  private computeClientActivityRate() {
-    this.clientActivityRate = 0;
-    if (this.statusHistory.length > 1) {
-      const [latest, prev] = this.statusHistory;
-      const seconds = NanoToMilli(latest.updated_at.subtract(prev.updated_at).toNumber()) / 1000;
-      const totalOps = (latest.metrics["sql.select.count"] - prev.metrics["sql.select.count"]) +
-        (latest.metrics["sql.update.count"] - prev.metrics["sql.update.count"]) +
-        (latest.metrics["sql.insert.count"] - prev.metrics["sql.insert.count"]) +
-        (latest.metrics["sql.delete.count"] - prev.metrics["sql.delete.count"]);
-      this.clientActivityRate = totalOps / seconds;
-    }
-  }
-}
 
 interface NodeSimulatorProps {
   nodesSummary: NodesSummary;
@@ -123,7 +59,7 @@ interface NodeSimulatorOwnProps {
 // having to query time series data. This is a hack because this behavior
 // should be moved into the reducer or into a higher order component.
 class NodeSimulator extends React.Component<NodeSimulatorProps & NodeSimulatorOwnProps, any> {
-  nodeHistories: { [id: string]: SimulatedNodeStatus } = {};
+  nodeHistories: { [id: string]: NodeHistory } = {};
 
   // accumulateHistory parses incoming nodeStatus properties and accumulates
   // a history for each node.
@@ -135,7 +71,7 @@ class NodeSimulator extends React.Component<NodeSimulatorProps & NodeSimulatorOw
     props.nodesSummary.nodeStatuses.map((status) => {
       const id = status.desc.node_id;
       if (!this.nodeHistories.hasOwnProperty(id)) {
-        this.nodeHistories[id] = new SimulatedNodeStatus(status);
+        this.nodeHistories[id] = new NodeHistory(status);
       } else {
         this.nodeHistories[id].update(status);
       }
